importer/requirements.txt
mysql-connector-python
requests
